backend/requirements.txt
beautifulsoup4>=4.12,<5
duckduckgo-search>=6.1,<7
fastapi>=0.115,<1
groq>=0.9,<1
httpx>=0.27,<1
pydantic>=2.9,<3
pytest>=8.3,<9
python-dotenv>=1.0,<2
uvicorn[standard]>=0.30,<1

